Rename to choose-assignee
