+++ lib/metadata/index.js
@@ -9,7 +9,7 @@ var filenames = require('./filenames'),
     status = require('./status'),
     labels = require('./labels'),
     getReviewers = require('./get-reviewers'),
-    getAssignee = require('./get-assignee'),
+    chooseAssignee = require('./choose-assignee'),
     funk = require('../funk'),
     github = require('../github');
 
@@ -105,7 +105,7 @@ module.exports = function getMetadada(number, author, content) {
 
             metadata.missingReviewers = getReviewers(metadata);
 
-            return getAssignee(number, metadata).then(function(a) {
+            return chooseAssignee(number, metadata).then(function(a) {
                 metadata.missingAssignee = a;
                 return metadata;
             });
